Remove background images, let elements fall naturally on page.
The positioning of the background images in the header and footer were a
recurring problem (and anyway they kinda looked like stock images for a
WordPress theme). I also don't think the footer is important enough that
it needs to *always* be on screen, if the user's display is small enough
to require scrolling. So, this way makes it easier to avoid having the
footer image appearing too high, or having various elements overlap each
other.